split server side js
